# LW OpenPose: Added torch.set_flush_denormal(True) for major increase in cpu performance